% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_ts.R
\name{compile_ts}
\alias{compile_ts}
\title{Compile TypeScript definitions}
\usage{
compile_ts(file, output = sprintf("\%s.d.ts", tools::file_path_sans_ext(file)))
}
\arguments{
\item{file}{the file containing marked-up R functions}

\item{output}{the location to save ts type files}
}
\value{
null, called for sideeffect of writing a file
}
\description{
Compile TypeScript definitions
}
