% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_dataframe}
\alias{ts_dataframe}
\title{Typed dataframe}
\usage{
ts_dataframe(...)
}
\arguments{
\item{...}{Named types.}
}
\value{
A ts object that accepts data frames with the specified types.
}
\description{
This is essentially a list, but the elements must have names and are all the same length.
}
