% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_null}
\alias{ts_null}
\title{Null type}
\usage{
ts_null()
}
\value{
A ts object that only accepts \code{NULL}.
}
\description{
This is a type that only accepts \code{NULL}. For function return types, use \code{ts_void()}.
}
\examples{
x <- ts_null()
x$check(NULL)
}
