% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_optional}
\alias{ts_optional}
\title{Optional type}
\usage{
ts_optional(type)
}
\arguments{
\item{type}{Type that is optional}
}
\value{
A \code{ts_object} that accepts either \code{type} or undefined.
}
\description{
A wrapper around union of a type and undefined
}
\examples{
x <- ts_optional(ts_numeric(1))
}
