% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_self}
\alias{ts_self}
\title{Self object}
\usage{
ts_self(n = -1L)
}
\arguments{
\item{n}{number of elements, either n = 1 (for singular) or
n != 0 (for array)}
}
\value{
a representation of 'self'
}
\description{
Representation of the 'self' property, used by recursive list definitions
}
