% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_void}
\alias{ts_void}
\title{Void type}
\usage{
ts_void()
}
\value{
A ts object that accepts \code{NULL}.
}
\description{
This is a type that accepts null values (this would typically be used for
functions that return nothing).
}
\seealso{
\code{\link{ts_null}()}
}
