% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\docType{class}
\name{tsWidget-class}
\alias{tsWidget-class}
\alias{tsWidget}
\title{Base Widget Class}
\description{
Base reference class for all widgets created with \code{createWidget()}.
This class provides the core functionality for reactive properties and
state synchronization with 'TypeScript'.
}

