% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{create_child_connector}
\alias{create_child_connector}
\title{Create Child Widget Connector}
\usage{
create_child_connector(
  child_instance,
  parent_instance,
  property_name,
  type_info,
  widget_def
)
}
\arguments{
\item{child_instance}{The child widget instance}

\item{parent_instance}{The parent widget instance}

\item{property_name}{Name of the property containing the child}

\item{type_info}{Type information from the widget definition}

\item{widget_def}{The widget definition object}
}
\value{
A TypeScript function constructor for the child widget
}
\description{
Internal helper function to create connector functions for child widgets.
Used by the \code{add_child} method of \code{tsWidget}.
}
\keyword{internal}
