% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{js_function}
\alias{js_function}
\title{JS functions callable from R}
\usage{
js_function(..., result = NULL)
}
\arguments{
\item{...}{arguments passed to the function}

\item{result}{the type of value returned from JS to R}
}
\value{
A ts object that accepts js functions (as input).
Currently not able to pass as output (but should, in future ...).
}
\description{
If result is NULL, it will be an oobSend (R process will continue),
otherwise R process will wait for a response (oobMessage).
}
\details{
TODO: when compiling, automatically wrap in self.oobMessage() or self.oobSend(), as necessary... ?
\itemize{
\item how about naked js functions? i.e., we might want to pass a function \emph{back} to javascript, for some reason?
}
}
