% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{rts_debug_enabled}
\alias{rts_debug_enabled}
\title{Check if debug logging is enabled for a tag}
\usage{
rts_debug_enabled(tag = "general")
}
\arguments{
\item{tag}{Character tag to check}
}
\value{
Logical
}
\description{
Check if debug logging is enabled for a tag
}
\keyword{internal}
