% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{rts_log}
\alias{rts_log}
\title{Log a debug message}
\usage{
rts_log(..., tag = "general")
}
\arguments{
\item{...}{Message parts (passed to paste0)}

\item{tag}{Character tag for filtering}
}
\description{
Log a debug message
}
\keyword{internal}
