% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_array}
\alias{ts_array}
\title{Array type}
\usage{
ts_array(type)
}
\arguments{
\item{type}{The input type, either a 'zod'-style string ("z.number()") or a \code{ts_object}.}
}
\value{
For a \code{ts_object} input, a \code{ts_object} wrapping \code{z.array(...)}.
For a character 'Zod' fragment, a character string schema.
}
\description{
An array of typed objects. In 'zod', these are represented
by \code{z.array()}; returned objects must be R lists, \code{Robj.list()}.
}
\examples{
x <- ts_array(ts_numeric(1))
x$check(list(1, 2, 3))
}
