% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_array}
\alias{ts_array}
\title{Array type}
\usage{
ts_array(type)
}
\arguments{
\item{type}{The input type, either a zod-style string ("z.number()") or a ts_object.}
}
\value{
An array object
}
\description{
An array of typed objects. In zod, these are represented
by \code{z.array()}; returned objects must be R lists, \code{Robj.list()}.
}
