% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{ts_compile}
\alias{ts_compile}
\title{Compile R functions to TypeScript schemas}
\usage{
ts_compile(f, ..., name, file)
}
\arguments{
\item{f}{A function or file path}

\item{...}{Additional arguments}

\item{name}{The name of the function}

\item{file}{The file path to write the TypeScript schema (optional). If \code{""}, the output is printed to the standard output console (see \code{cat}).}
}
\value{
Character vector of TypeScript schema, or NULL if writing to file
}
\description{
Compile R functions to TypeScript schemas
}
