% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_dataframe}
\alias{ts_dataframe}
\title{Typed dataframe}
\usage{
ts_dataframe(..., default = NULL)
}
\arguments{
\item{...}{Named types.}
}
\value{
A ts object that accepts data frames with the specified types.
}
\description{
This is essentially a list, but the elements must have names and are all the same length.
}
\examples{
x <- ts_dataframe(a = ts_integer(1), b = ts_character(1))
x$check(data.frame(a = 1L, b = "a"))
}
