% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{ts_deploy}
\alias{ts_deploy}
\title{Deploy a ts Rserve app}
\usage{
ts_deploy(
  f,
  file = sprintf("\%s.rserve.R", tools::file_path_sans_ext(f)),
  init = NULL,
  port = 6311,
  run = c("no", "here", "background"),
  silent = FALSE
)
}
\arguments{
\item{f}{The path to the application files}

\item{file}{The file to write the deployment script to}

\item{init}{Names of objects (ts_functions) to make available to
the initialisation function}

\item{port}{The port to deploy the app on}

\item{run}{Whether to run the deployment script,
takes values "no", "here", "background"}

\item{silent}{Whether to print the deployment script}
}
\value{
NULL, called to open an Rserve instance
}
\description{
Deploy a ts Rserve app
}
