% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{ts_function}
\alias{ts_function}
\title{TS function definition}
\usage{
ts_function(f, ..., result = ts_void())
}
\arguments{
\item{f}{an R function}

\item{...}{argument definitions (only required if f does not specify these in its formals)}

\item{result}{return type (ignored if overloads are provided)}
}
\description{
TS function definition
}
