% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_list}
\alias{ts_list}
\title{Typed list}
\usage{
ts_list(...)
}
\arguments{
\item{...}{A list of types, named or unnamed.}
}
\value{
A ts object that accepts lists with the specified types.
}
\description{
A list is a vector of other robjects, which may or may not be named.
}
