% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_union}
\alias{ts_union}
\title{Union type}
\usage{
ts_union(..., default = NULL)
}
\arguments{
\item{...}{Type objects to merge}

\item{default}{Default value for the type (optional).}
}
\value{
A \code{ts_object} representing a 'Zod' union of the given types.
}
\description{
Create a union of types. Currently this only accepts schemas as strings.
}
\examples{
x <- ts_union(ts_numeric(1), ts_character(1))
}
