% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ts_union}
\alias{ts_union}
\title{Union type}
\usage{
ts_union(...)
}
\arguments{
\item{...}{Zod types to merge (as strings)}
}
\description{
Create a union of types. Currently this only accepts schemas as strings.
}
\examples{
x <- ts_union("z.number()", "z.string()")
}
